separate common code
