+++ pyiron_contrib/__init__.py
@@ -47,8 +47,8 @@ JOB_CLASS_DICT['StorageJob'] = 'pyiron_contrib.RDM.storagejob'
 JOB_CLASS_DICT['PacemakerJob'] = 'pyiron_contrib.atomistics.pacemaker.job'
 JOB_CLASS_DICT['MeamFit'] = 'pyiron_contrib.atomistics.meamfit.meamfit'
 JOB_CLASS_DICT['Cp2kJob'] = 'pyiron_contrib.atomistics.cp2k.job'
-JOB_CLASS_DICT['Piglet'] = 'pyiron_contrib.atomistics.pimd.piglet'
-JOB_CLASS_DICT['PigletMlip'] = 'pyiron_contrib.atomistics.pimd.piglet_mlip'
+JOB_CLASS_DICT['Piglet'] = 'pyiron_contrib.atomistics.pimd.pimd_jobs'
+JOB_CLASS_DICT['PigletMlip'] = 'pyiron_contrib.atomistics.pimd.pimd_jobs'
 
 
 from ._version import get_versions
